Make game setup fluent
